Header component using Search component
