+++ src/components/header/Header.jsx
@@ -3,12 +3,12 @@ import Navigation from "../navigation/Navigation";
 import Logo from "../logo/Logo";
 import { FaUser } from "react-icons/fa";
 import { IoMdCart } from "react-icons/io";
-import { FaSearch } from "react-icons/fa";
 import { GiHamburgerMenu } from "react-icons/gi";
 import { IoClose } from "react-icons/io5";
 import ShoppingCart from './shoppingcart/ShoppingCart';
 import { useCart } from './shoppingcart/CartContext';
-import { Link, useLoaderData } from 'react-router';
+// import { Link, useLoaderData } from 'react-router';
+import Search from '../Search';
 
 export default function Header() {
     const [showMenu, setShowMenu] = useState(false);
@@ -28,15 +28,10 @@ export default function Header() {
         setActive(!active);
     };
 
-    // search functionality
-    const searchData = useLoaderData();
-    const [search, setSearch] = useState("");
-    console.log(searchData);
     const handleChange = e => {
         setSearch(e.target.value)
     }
-    console.log(search);
-    
+    // console.log(search);    
 
     return (
         <header className="header">
@@ -45,21 +40,7 @@ export default function Header() {
                 <Navigation />
             </div>
             <div className="header__searchUserCart">
-                <div className="header__search">
-                    <input type="text" name="search" id="search" placeholder="Search product..." onChange={handleChange} onClick={handleClick} />
-                    <FaSearch
-                        onClick={handleClick}
-                        style={{ color: active ? 'red' : 'white', cursor: 'pointer' }} />
-                    <div className="searchShow">
-                        {active && searchData && searchData
-                            .filter((itm) =>
-                                search.trim() !== '' && itm.type.toLowerCase().includes(search.toLowerCase())
-                            )
-                            .map((itm) => (
-                                <Link to={`/products/${itm.id}`} key={itm.id}>{itm.type}</Link>
-                            ))}
-                    </div>
-                </div>
+                <Search />
                 <div className="header__user">
                     <FaUser color="white" />
                 </div>
@@ -76,20 +57,13 @@ export default function Header() {
             </div>
             {showMenu && (
                 <>
-                    <IoClose color="red" onClick={() => setShowMenu(false)} className="closeIcon" />
-                    {/* <GiHamburgerMenu color="white" className="hamburger" onClick={() => setShowMenu(false)}/>                 */}
+                    <IoClose color="orange" onClick={() => setShowMenu(false)} className="closeIcon" />
                     <div className="header--mobile">
                         <Logo />
                         <Navigation className="mobile--navigation" />
                     </div>
                 </>
             )}
-
-            {/* {active && (
-                <>
-                    <input type="text" name="search" id="search" placeholder="Search product..." />
-                </>
-            )} */}
         </header>
     )
 }
